% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KPI.R
\name{calc_kpi}
\alias{calc_kpi}
\title{Calculate all KPI metrics}
\usage{
calc_kpi(file_path)
}
\arguments{
\item{file_path}{path to Excel file}
}
\value{
data.frame of KPI metrics
}
\description{
Main function that processes a file path and returns all KPIs.
}
